requirements.txt
google-cloud-bigquery==2.12.0
google-cloud-storage==1.41.0
google-cloud-logging==2.6.0